separate convert tag ids
